// ombi-api/model/ombi-store-repository-requests-episode-requests.ts
// tslint:disable
/// <reference path="../custom.d.ts" />
/**
 * Ombi Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { OmbiStoreRepositoryRequestsSeasonRequests } from './ombi-store-repository-requests-season-requests';

/**
 * 
 * @export
 * @interface OmbiStoreRepositoryRequestsEpisodeRequests
 */
export interface OmbiStoreRepositoryRequestsEpisodeRequests {
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    episodeNumber?: number;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    title?: string;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    airDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    url?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    available?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    approved?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    requested?: boolean;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    seasonId?: number;
    /**
     * 
     * @type {OmbiStoreRepositoryRequestsSeasonRequests}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    season?: OmbiStoreRepositoryRequestsSeasonRequests;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    airDateDisplay?: string;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreRepositoryRequestsEpisodeRequests
     */
    id?: number;
}