// ombi-api/model/ombi-store-entities-requests-tv-requests.ts
// tslint:disable
/// <reference path="../custom.d.ts" />
/**
 * Ombi Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { OmbiStoreEntitiesRequestsChildRequests } from './ombi-store-entities-requests-child-requests';

/**
 * 
 * @export
 * @interface OmbiStoreEntitiesRequestsTvRequests
 */
export interface OmbiStoreEntitiesRequestsTvRequests {
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    tvDbId?: number;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    imdbId?: string;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    qualityOverride?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    rootFolder?: number;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    overview?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    posterPath?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    background?: string;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    releaseDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    status?: string;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    totalSeasons?: number;
    /**
     * 
     * @type {Array<OmbiStoreEntitiesRequestsChildRequests>}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    childRequests?: Array<OmbiStoreEntitiesRequestsChildRequests>;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsTvRequests
     */
    id?: number;
}